Fixed: removed x64 warning.
git-svn-id: https://svn.code.sf.net/p/akelpad/codesvn/trunk@3951 24cdac72-59cb-4434-bdd7-86cdcaa6fc6e

+++ akelpad_4/AkelEdit/AkelEdit.c
@@ -1324,9 +1324,9 @@ LRESULT CALLBACK AE_EditProc(AKELEDIT *ae, UINT uMsg, WPARAM wParam, LPARAM lPar
     case AEM_VPOSFROMLINE:
     {
       if (wParam == AECT_GLOBAL)
-        return AE_VPosFromLine(ae, lParam);
+        return AE_VPosFromLine(ae, (int)lParam);
       if (wParam == AECT_CLIENT)
-        return ae->rcDraw.top + (AE_VPosFromLine(ae, lParam) - ae->nVScrollPos);
+        return ae->rcDraw.top + (AE_VPosFromLine(ae, (int)lParam) - ae->nVScrollPos);
     }
     case AEM_GETMOUSESTATE:
     {
